+++ include/obake/polynomials/packed_monomial.hpp
@@ -24,9 +24,6 @@
 
 #include <boost/serialization/access.hpp>
 
-#include <fmt/format.h>
-#include <fmt/ostream.h>
-
 #include <tbb/blocked_range.h>
 #include <tbb/parallel_invoke.h>
 #include <tbb/parallel_reduce.h>
@@ -73,9 +70,8 @@ public:
     constexpr explicit packed_monomial(const T &n) : m_value(n) {}
     // Constructor from input iterator and size.
     template <typename It>
-    requires InputIterator<It> &&
-        SafelyCastable<typename ::std::iterator_traits<It>::reference, T> constexpr explicit packed_monomial(It it,
-                                                                                                             unsigned n)
+        requires InputIterator<It> && SafelyCastable<typename ::std::iterator_traits<It>::reference, T>
+    constexpr explicit packed_monomial(It it, unsigned n)
     {
         kpacker<T> kp(n);
         for (auto i = 0u; i < n; ++i, ++it) {
@@ -100,24 +96,25 @@ private:
 public:
     // Ctor from a pair of forward iterators.
     template <typename It>
-    requires ForwardIterator<It> &&SafelyCastable<typename ::std::iterator_traits<It>::difference_type, unsigned> &&
-        SafelyCastable<typename ::std::iterator_traits<It>::reference, T> constexpr explicit packed_monomial(It b, It e)
-        : packed_monomial(fwd_it_ctor_tag{}, b, e)
+        requires ForwardIterator<It> && SafelyCastable<typename ::std::iterator_traits<It>::difference_type, unsigned>
+                 && SafelyCastable<typename ::std::iterator_traits<It>::reference, T>
+    constexpr explicit packed_monomial(It b, It e) : packed_monomial(fwd_it_ctor_tag{}, b, e)
     {
     }
     // Ctor from forward range.
     template <typename Range>
-    requires ForwardRange<Range> &&
-        SafelyCastable<typename ::std::iterator_traits<range_begin_t<Range>>::difference_type, unsigned> &&
-            SafelyCastable<typename ::std::iterator_traits<range_begin_t<Range>>::reference,
-                           T> constexpr explicit packed_monomial(Range &&r)
+        requires ForwardRange<Range>
+                 && SafelyCastable<typename ::std::iterator_traits<range_begin_t<Range>>::difference_type, unsigned>
+                 && SafelyCastable<typename ::std::iterator_traits<range_begin_t<Range>>::reference, T>
+    constexpr explicit packed_monomial(Range &&r)
         : packed_monomial(fwd_it_ctor_tag{}, ::obake::begin(::std::forward<Range>(r)),
                           ::obake::end(::std::forward<Range>(r)))
     {
     }
     // Ctor from init list.
     template <typename U>
-    requires SafelyCastable<const U &, T> constexpr explicit packed_monomial(::std::initializer_list<U> l)
+        requires SafelyCastable<const U &, T>
+    constexpr explicit packed_monomial(::std::initializer_list<U> l)
         : packed_monomial(fwd_it_ctor_tag{}, l.begin(), l.end())
     {
     }
@@ -290,10 +287,11 @@ inline constexpr bool same_packed_monomial_v = same_packed_monomial<T, U>::value
 //   the constraints on packed_monomial),
 // - the random-access iterator concept.
 template <typename R1, typename R2>
-requires InputRange<R1> &&InputRange<R2> &&detail::same_packed_monomial_v<
-    remove_cvref_t<typename ::std::iterator_traits<range_begin_t<R1>>::reference>,
-    remove_cvref_t<typename ::std::iterator_traits<range_begin_t<R2>>::reference>> inline bool
-monomial_range_overflow_check(R1 &&r1, R2 &&r2, const symbol_set &ss)
+    requires InputRange<R1> && InputRange<R2>
+             && detail::same_packed_monomial_v<
+                 remove_cvref_t<typename ::std::iterator_traits<range_begin_t<R1>>::reference>,
+                 remove_cvref_t<typename ::std::iterator_traits<range_begin_t<R2>>::reference>>
+inline bool monomial_range_overflow_check(R1 &&r1, R2 &&r2, const symbol_set &ss)
 {
     using pm_t = remove_cvref_t<typename ::std::iterator_traits<range_begin_t<R1>>::reference>;
     using value_type = typename pm_t::value_type;
@@ -520,7 +518,8 @@ OBAKE_DLL_PUBLIC ::std::uint64_t key_p_degree(const packed_monomial<::std::uint6
 template <typename T, typename U,
           ::std::enable_if_t<::std::disjunction_v<::obake::detail::is_mppp_integer<U>,
                                                   is_safely_convertible<const U &, ::mppp::integer<1> &>>,
-                             int> = 0>
+                             int>
+          = 0>
 inline packed_monomial<T> monomial_pow(const packed_monomial<T> &p, const U &n, const symbol_set &ss)
 {
     assert(polynomials::key_is_compatible(p, ss));
@@ -537,16 +536,8 @@ inline packed_monomial<T> monomial_pow(const packed_monomial<T> &p, const U &n, 
             ::mppp::integer<1> ret;
 
             if (obake_unlikely(!::obake::safe_convert(ret, n))) {
-                if constexpr (is_stream_insertable_v<const U &>) {
-                    // Provide better error message if U is ostreamable.
-                    using namespace ::fmt::literals;
-                    obake_throw(::std::invalid_argument,
-                                "Invalid exponent for monomial exponentiation: the exponent ({}) "
-                                "cannot be converted into an integral value"_format(n));
-                } else {
-                    obake_throw(::std::invalid_argument, "Invalid exponent for monomial exponentiation: the exponent "
-                                                         "cannot be converted into an integral value");
-                }
+                obake_throw(::std::invalid_argument, "Invalid exponent for monomial exponentiation: the exponent "
+                                                     "cannot be converted into an integral value");
             }
 
             return ret;